+++ TigoWeb/urlpattern.go
@@ -84,28 +84,6 @@ func (urlPattern *UrlPattern) AppendRouterPattern(pattern Pattern, v interface {
 	urlPattern.router.HandleFunc(pattern.Url, middlewares(v.Handle))
 }
 
-func convertHandleFuncMV(v Middleware) middleware {
-	return func(next http.HandlerFunc) http.HandlerFunc {
-		return func(w http.ResponseWriter, r *http.Request) {
-			// 此处需要判断请求是否继续交给下一个中间件处理
-			if isGoOn := v(&w, r); isGoOn {
-				next.ServeHTTP(w, r)
-			}
-		}
-	}
-}
-
-func convertHandleMV(v Middleware) mux.MiddlewareFunc {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			// 此处需要判断请求是否继续交给下一个中间件处理
-			if isGoOn := v(&w, r); isGoOn {
-				next.ServeHTTP(w, r)
-			}
-		})
-	}
-}
-
 // Init 初始化url映射，遍历UrlMapping，将handler与对应的URL依次挂载到http服务上
 func (urlPattern *UrlPattern) Init() {
 	for _, pattern := range urlPattern.UrlPatterns {
